split out top toolbar. TODO: figure out why tda tab is stuck
